requirements.txt
frida-tools
psutil
pywin32
pefile